Reset branch to more recent main. Reorganised gesture folders again
