+++ instrumentation/view-click/src/main/kotlin/io/opentelemetry/android/instrumentation/view/click/ViewClickEventGenerator.kt
@@ -8,17 +8,15 @@ package io.opentelemetry.android.instrumentation.view.click
 import android.view.GestureDetector
 import android.view.MotionEvent
 import android.view.View
-import android.view.ViewGroup
 import android.view.Window
-import io.opentelemetry.android.instrumentation.view.click.internal.APP_SCREEN_CLICK_EVENT_NAME
-import io.opentelemetry.android.instrumentation.view.click.internal.Gesture
-import io.opentelemetry.android.instrumentation.view.click.internal.VIEW_CLICK_EVENT_NAME
-import io.opentelemetry.android.semconv.events.AppScreenFlingEvent
+import io.opentelemetry.android.instrumentation.view.common.Gesture
+import io.opentelemetry.android.instrumentation.view.common.ToggleableTracker
+import io.opentelemetry.android.instrumentation.view.common.TouchEventConsumer
+import io.opentelemetry.android.instrumentation.view.common.WindowCallbackWrapper
+import io.opentelemetry.android.instrumentation.view.common.findTargetForTap
+import io.opentelemetry.android.instrumentation.view.common.viewToName
 import io.opentelemetry.android.semconv.events.AppScreenLongpressEvent
-import io.opentelemetry.android.semconv.events.AppScreenScrollEvent
-import io.opentelemetry.android.semconv.events.AppWidgetFlingEvent
 import io.opentelemetry.android.semconv.events.AppWidgetLongpressEvent
-import io.opentelemetry.android.semconv.events.AppWidgetScrollEvent
 import io.opentelemetry.api.common.Attributes
 import io.opentelemetry.api.logs.LogRecordBuilder
 import io.opentelemetry.api.logs.Logger
@@ -28,11 +26,11 @@ import io.opentelemetry.kotlin.semconv.AppAttributes.APP_WIDGET_ID
 import io.opentelemetry.kotlin.semconv.AppAttributes.APP_WIDGET_NAME
 import io.opentelemetry.kotlin.semconv.IncubatingApi
 import java.lang.ref.WeakReference
-import java.util.LinkedList
 
 internal class ViewClickEventGenerator(
     private val eventLogger: Logger,
-) {
+) : TouchEventConsumer,
+    ToggleableTracker {
     private var windowRef: WeakReference<Window>? = null
 
     @OptIn(IncubatingApi::class)
@@ -49,7 +47,7 @@ internal class ViewClickEventGenerator(
                         .setAttribute(APP_SCREEN_COORDINATE_X, safeEvent.x.toLong())
                         .emit()
 
-                    findTargetForTap(window.decorView, safeEvent.x, safeEvent.y)?.let { view ->
+                    findTargetForTap(window.decorView, safeEvent.x, safeEvent.y, viewCoordinates)?.let { view ->
                         createEvent(VIEW_CLICK_EVENT_NAME, gesture)
                             .setAllAttributes(createViewAttributes(view))
                             .emit()
@@ -71,7 +69,7 @@ internal class ViewClickEventGenerator(
                         .setAttribute(APP_SCREEN_COORDINATE_X, safeEvent.x.toLong())
                         .emit()
 
-                    findTargetForTap(window.decorView, safeEvent.x, safeEvent.y)?.let { view ->
+                    findTargetForTap(window.decorView, safeEvent.x, safeEvent.y, viewCoordinates)?.let { view ->
                         createEvent(VIEW_CLICK_EVENT_NAME, gesture)
                             .setAllAttributes(createViewAttributes(view))
                             .emit()
@@ -88,51 +86,11 @@ internal class ViewClickEventGenerator(
                     val gesture = Gesture.LongPress(safeEvent)
                     emitAppScreenLongPress(safeEvent, gesture)
 
-                    findTargetForTap(window.decorView, safeEvent.x, safeEvent.y)?.let { view ->
+                    findTargetForTap(window.decorView, safeEvent.x, safeEvent.y, viewCoordinates)?.let { view ->
                         emitAppWidgetLongPress(view, gesture)
                     }
                     safeEvent.recycle()
                 }
-            }
-
-            override fun onScroll(
-                e1: MotionEvent?,
-                e2: MotionEvent,
-                distanceX: Float,
-                distanceY: Float,
-            ): Boolean {
-                windowRef?.get()?.let { window ->
-
-                    val safeEvent = MotionEvent.obtain(e2)
-                    val gesture = Gesture.Scroll(safeEvent, distanceX.toDouble(), distanceY.toDouble())
-                    emitAppScreenScroll(safeEvent, gesture)
-
-                    findTargetForTap(window.decorView, safeEvent.x, safeEvent.y)?.let { view ->
-                        emitAppWidgetScroll(view, gesture)
-                    }
-                    safeEvent.recycle()
-                }
-                return false
-            }
-
-            override fun onFling(
-                e1: MotionEvent?,
-                e2: MotionEvent,
-                velocityX: Float,
-                velocityY: Float,
-            ): Boolean {
-                windowRef?.get()?.let { window ->
-
-                    val safeEvent = MotionEvent.obtain(e2)
-                    val gesture = Gesture.Fling(safeEvent, velocityX.toDouble(), velocityY.toDouble())
-                    emitAppScreenFling(safeEvent, gesture)
-
-                    findTargetForTap(window.decorView, safeEvent.x, safeEvent.y)?.let { view ->
-                        emitAppWidgetFling(view, gesture)
-                    }
-                    safeEvent.recycle()
-                }
-                return false
             }
         }
 
@@ -140,19 +98,19 @@ internal class ViewClickEventGenerator(
 
     private val viewCoordinates = IntArray(2)
 
-    fun startTracking(window: Window) {
+    override fun startTracking(window: Window) {
         windowRef = WeakReference(window)
         val currentCallback = window.callback
         window.callback = WindowCallbackWrapper(currentCallback, this)
     }
 
-    fun generateClick(motionEvent: MotionEvent?) {
-        if (motionEvent != null) {
-            gestureDetector.onTouchEvent(motionEvent)
+    override fun consumeEvent(event: MotionEvent?) {
+        if (event != null) {
+            gestureDetector.onTouchEvent(event)
         }
     }
 
-    fun stopTracking() {
+    override fun stopTracking() {
         windowRef?.get()?.run {
             if (callback is WindowCallbackWrapper) {
                 callback = (callback as WindowCallbackWrapper).unwrap()
@@ -200,66 +158,6 @@ internal class ViewClickEventGenerator(
         ).emit(eventLogger)
     }
 
-    private fun emitAppScreenScroll(
-        motionEvent: MotionEvent,
-        gesture: Gesture.Scroll,
-    ) {
-        AppScreenScrollEvent(
-            appScreenCoordinateX = motionEvent.x.toLong(),
-            appScreenCoordinateY = motionEvent.y.toLong(),
-            hwPointerButton = gesture.buttonStateDescription,
-            hwPointerDistanceX = gesture.distanceX,
-            hwPointerDistanceY = gesture.distanceY,
-            hwPointerType = gesture.toolTypeDescription,
-        ).emit(eventLogger)
-    }
-
-    private fun emitAppWidgetScroll(
-        view: View,
-        gesture: Gesture.Scroll,
-    ) {
-        AppWidgetScrollEvent(
-            appScreenCoordinateX = view.x.toLong(),
-            appScreenCoordinateY = view.y.toLong(),
-            appWidgetId = view.id.toString(),
-            appWidgetName = viewToName(view),
-            hwPointerButton = gesture.buttonStateDescription,
-            hwPointerDistanceX = gesture.distanceX,
-            hwPointerDistanceY = gesture.distanceY,
-            hwPointerType = gesture.toolTypeDescription,
-        ).emit(eventLogger)
-    }
-
-    private fun emitAppScreenFling(
-        motionEvent: MotionEvent,
-        gesture: Gesture.Fling,
-    ) {
-        AppScreenFlingEvent(
-            appScreenCoordinateX = motionEvent.x.toLong(),
-            appScreenCoordinateY = motionEvent.y.toLong(),
-            hwPointerButton = gesture.buttonStateDescription,
-            hwPointerType = gesture.toolTypeDescription,
-            hwPointerVelocityX = gesture.velocityX,
-            hwPointerVelocityY = gesture.velocityY,
-        ).emit(eventLogger)
-    }
-
-    private fun emitAppWidgetFling(
-        view: View,
-        gesture: Gesture.Fling,
-    ) {
-        AppWidgetFlingEvent(
-            appScreenCoordinateX = view.x.toLong(),
-            appScreenCoordinateY = view.y.toLong(),
-            appWidgetId = view.id.toString(),
-            appWidgetName = viewToName(view),
-            hwPointerButton = gesture.buttonStateDescription,
-            hwPointerType = gesture.toolTypeDescription,
-            hwPointerVelocityX = gesture.velocityX,
-            hwPointerVelocityY = gesture.velocityY,
-        ).emit(eventLogger)
-    }
-
     @OptIn(IncubatingApi::class)
     private fun createViewAttributes(view: View): Attributes {
         val builder = Attributes.builder()
@@ -270,74 +168,7 @@ internal class ViewClickEventGenerator(
         builder.put(APP_SCREEN_COORDINATE_Y, view.y.toLong())
         return builder.build()
     }
-
-    private fun viewToName(view: View): String =
-        try {
-            view.resources?.getResourceEntryName(view.id) ?: view.id.toString()
-        } catch (throwable: Throwable) {
-            view.id.toString()
-        }
-
-    private fun findTargetForTap(
-        decorView: View,
-        x: Float,
-        y: Float,
-    ): View? {
-        val queue = LinkedList<View>()
-        queue.addFirst(decorView)
-        var target: View? = null
-
-        while (queue.isNotEmpty()) {
-            val view = queue.removeFirst()
-            if (isJetpackComposeView(view)) {
-                return null
-            }
-
-            if (isValidClickTarget(view)) {
-                target = view
-            }
-
-            if (view is ViewGroup) {
-                handleViewGroup(view, x, y, queue)
-            }
-        }
-        return target
-    }
-
-    private fun isValidClickTarget(view: View): Boolean = view.isClickable && view.isVisible
-
-    private fun handleViewGroup(
-        view: ViewGroup,
-        x: Float,
-        y: Float,
-        stack: LinkedList<View>,
-    ) {
-        if (!view.isVisible) return
-
-        for (i in 0 until view.childCount) {
-            val child = view.getChildAt(i)
-            if (hitTest(child, x, y) && !isJetpackComposeView(child)) {
-                stack.add(child)
-            }
-        }
-    }
-
-    private fun hitTest(
-        view: View,
-        x: Float,
-        y: Float,
-    ): Boolean {
-        view.getLocationInWindow(viewCoordinates)
-        val vx = viewCoordinates[0]
-        val vy = viewCoordinates[1]
-
-        val w = view.width
-        val h = view.height
-        return !(x < vx || x > vx + w || y < vy || y > vy + h)
-    }
-
-    private fun isJetpackComposeView(view: View): Boolean = view::class.java.name.startsWith("androidx.compose.ui.platform.ComposeView")
-
-    private val View.isVisible: Boolean
-        get() = visibility == View.VISIBLE
 }
+
+internal const val APP_SCREEN_CLICK_EVENT_NAME = "app.screen.click"
+internal const val VIEW_CLICK_EVENT_NAME = "app.widget.click"
