+++ instrumentation/view-common/README.md
@@ -1,0 +1,3 @@
+# Common Module
+
+For shared code between the view-based gesture instrumentations
