+++ instrumentation/view-common/build.gradle.kts
@@ -1,0 +1,22 @@
+plugins {
+    id("otel.android-library-conventions")
+}
+
+description = "OpenTelemetry Android View Gesture common utils"
+
+android {
+    namespace = "io.opentelemetry.android.instrumentation.view.common"
+
+    defaultConfig {
+        consumerProguardFiles("consumer-rules.pro")
+    }
+}
+
+dependencies {
+    implementation(project(":agent-api"))
+    implementation(project(":semconv"))
+    implementation(project(":services"))
+    implementation(libs.opentelemetry.sdk)
+    implementation(libs.opentelemetry.instrumentation.api)
+    testImplementation(libs.androidx.fragment)
+}
