+++ instrumentation/view-common/src/main/kotlin/io/opentelemetry/android/instrumentation/view/common/ViewClickActivityCallback.kt
@@ -3,21 +3,28 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android.instrumentation.view.click
+package io.opentelemetry.android.instrumentation.view.common
 
 import android.app.Activity
+import android.view.Window
 import io.opentelemetry.android.internal.services.visiblescreen.activities.DefaultingActivityLifecycleCallbacks
 
-internal class ViewClickActivityCallback(
-    private val viewClickEventGenerator: ViewClickEventGenerator,
+class ViewClickActivityCallback(
+    private val toggleableTracker: ToggleableTracker,
 ) : DefaultingActivityLifecycleCallbacks {
     override fun onActivityResumed(activity: Activity) {
         super.onActivityResumed(activity)
-        viewClickEventGenerator.startTracking(activity.window)
+        toggleableTracker.startTracking(activity.window)
     }
 
     override fun onActivityPaused(activity: Activity) {
         super.onActivityPaused(activity)
-        viewClickEventGenerator.stopTracking()
+        toggleableTracker.stopTracking()
     }
+}
+
+interface ToggleableTracker {
+    fun startTracking(window: Window)
+
+    fun stopTracking()
 }
