+++ instrumentation/view-scroll/api/view-scroll.api
@@ -1,0 +1,6 @@
+public final class io/opentelemetry/android/instrumentation/view/scroll/ViewScrollInstrumentation : io/opentelemetry/android/instrumentation/AndroidInstrumentation {
+	public fun <init> ()V
+	public fun getName ()Ljava/lang/String;
+	public fun install (Landroid/content/Context;Lio/opentelemetry/android/OpenTelemetryRum;)V
+}
+
