+++ instrumentation/view-scroll/build.gradle.kts
@@ -3,10 +3,10 @@ plugins {
     id("otel.publish-conventions")
 }
 
-description = "OpenTelemetry Android View click library instrumentation"
+description = "OpenTelemetry Android View scroll library instrumentation"
 
 android {
-    namespace = "io.opentelemetry.android.instrumentation.view.click"
+    namespace = "io.opentelemetry.android.instrumentation.view.scroll"
 
     defaultConfig {
         consumerProguardFiles("consumer-rules.pro")
@@ -18,8 +18,8 @@ dependencies {
     implementation(project(":services"))
     implementation(project(":agent-api"))
     implementation(project(":instrumentation:android-instrumentation"))
-    implementation(project(":semconv"))
     implementation(project(":instrumentation:view-common"))
+    implementation(project(":semconv"))
 
     implementation(libs.opentelemetry.instrumentation.apiSemconv)
     implementation(libs.opentelemetry.semconv.kotlin)
