+++ instrumentation/view-scroll/consumer-rules.pro
@@ -1,0 +1,2 @@
+# used in reflection to check if compose is available at runtime
+-keepnames class androidx.compose.ui.platform.ComposeView
