+++ instrumentation/view-scroll/src/main/kotlin/io/opentelemetry/android/instrumentation/view/scroll/ViewScrollInstrumentation.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android.instrumentation.view.click
+package io.opentelemetry.android.instrumentation.view.scroll
 
 import android.app.Application
 import android.content.Context
@@ -13,8 +13,8 @@ import io.opentelemetry.android.instrumentation.AndroidInstrumentation
 import io.opentelemetry.android.instrumentation.view.common.ViewClickActivityCallback
 
 @AutoService(AndroidInstrumentation::class)
-class ViewClickInstrumentation : AndroidInstrumentation {
-    override val name: String = "view.click"
+class ViewScrollInstrumentation : AndroidInstrumentation {
+    override val name: String = "view.scroll"
 
     override fun install(
         context: Context,
@@ -22,10 +22,10 @@ class ViewClickInstrumentation : AndroidInstrumentation {
     ) {
         (context as? Application)?.registerActivityLifecycleCallbacks(
             ViewClickActivityCallback(
-                ViewClickEventGenerator(
+                ViewScrollEventGenerator(
                     openTelemetryRum.openTelemetry
                         .logsBridge
-                        .loggerBuilder("io.opentelemetry.android.instrumentation.view.click")
+                        .loggerBuilder("io.opentelemetry.android.instrumentation.view.scroll")
                         .build(),
                 ),
             ),
