+++ instrumentation/view-scroll/src/test/kotlin/io/opentelemetry/android/instrumentation/view/scroll/ViewScrollInstrumentationTest.kt
@@ -5,7 +5,7 @@
 
 @file:OptIn(IncubatingApi::class)
 
-package io.opentelemetry.android.instrumentation.view.click
+package io.opentelemetry.android.instrumentation.view.scroll
 
 import android.app.Activity
 import android.app.Application
@@ -19,11 +19,12 @@ import getScrollSequence
 import io.mockk.MockKAnnotations
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
-import io.mockk.junit5.MockKExtension
 import io.mockk.mockk
 import io.mockk.slot
 import io.mockk.verify
 import io.opentelemetry.android.OpenTelemetryRum
+import io.opentelemetry.android.instrumentation.view.common.ViewClickActivityCallback
+import io.opentelemetry.android.instrumentation.view.common.WindowCallbackWrapper
 import io.opentelemetry.android.semconv.HwAttributes.HW_POINTER_DISTANCE_X_KEY
 import io.opentelemetry.android.semconv.HwAttributes.HW_POINTER_DISTANCE_Y_KEY
 import io.opentelemetry.android.semconv.HwAttributes.HW_POINTER_TYPE_KEY
@@ -44,7 +45,6 @@ import io.opentelemetry.sdk.testing.junit4.OpenTelemetryRule
 import mockView
 import org.junit.Before
 import org.junit.Test
-import org.junit.jupiter.api.extension.ExtendWith
 import org.junit.runner.RunWith
 
 private const val APP_SCREEN_FLING_EVENT_NAME = "app.screen.fling"
@@ -53,7 +53,6 @@ private const val VIEW_FLING_EVENT_NAME = "app.widget.fling"
 private const val VIEW_SCROLL_EVENT_NAME = "app.widget.scroll"
 
 @RunWith(AndroidJUnit4::class)
-@ExtendWith(MockKExtension::class)
 class ViewScrollInstrumentationTest {
     private lateinit var openTelemetryRule: OpenTelemetryRule
 
@@ -93,7 +92,7 @@ class ViewScrollInstrumentationTest {
         every { activity.window } returns window
         every { application.registerActivityLifecycleCallbacks(any()) } returns Unit
 
-        ViewClickInstrumentation().install(application, openTelemetryRum)
+        ViewScrollInstrumentation().install(application, openTelemetryRum)
 
         verify {
             application.registerActivityLifecycleCallbacks(capture(callbackCapturingSlot))
@@ -166,7 +165,7 @@ class ViewScrollInstrumentationTest {
         every { activity.window } returns window
         every { application.registerActivityLifecycleCallbacks(any()) } returns Unit
 
-        ViewClickInstrumentation().install(application, openTelemetryRum)
+        ViewScrollInstrumentation().install(application, openTelemetryRum)
 
         verify {
             application.registerActivityLifecycleCallbacks(capture(callbackCapturingSlot))
@@ -218,7 +217,7 @@ class ViewScrollInstrumentationTest {
         every { activity.window } returns window
         every { application.registerActivityLifecycleCallbacks(any()) } returns Unit
 
-        ViewClickInstrumentation().install(application, openTelemetryRum)
+        ViewScrollInstrumentation().install(application, openTelemetryRum)
 
         verify {
             application.registerActivityLifecycleCallbacks(capture(callbackCapturingSlot))
@@ -294,7 +293,7 @@ class ViewScrollInstrumentationTest {
         every { activity.window } returns window
         every { application.registerActivityLifecycleCallbacks(any()) } returns Unit
 
-        ViewClickInstrumentation().install(application, openTelemetryRum)
+        ViewScrollInstrumentation().install(application, openTelemetryRum)
 
         verify {
             application.registerActivityLifecycleCallbacks(capture(callbackCapturingSlot))
