+++ test-common/src/test/kotlin/TestUtilsTest.kt
@@ -3,8 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import org.assertj.core.api.Assertions.assertThat
-import org.assertj.core.api.Assertions.within
+import org.assertj.core.api.Assertions
 import org.junit.Test
 
 class TestUtilsTest {
@@ -13,8 +12,8 @@ class TestUtilsTest {
         val pointA = arrayOf(20f, 20f)
         val expectedPointB = arrayOf(40f, 20f)
         val (pointBx, pointBy) = getDestinationPoint(pointA, 20, 0.0)
-        assertThat(pointBx).isCloseTo(expectedPointB[0], within(1f))
-        assertThat(pointBy).isCloseTo(expectedPointB[1], within(1f))
+        Assertions.assertThat(pointBx).isCloseTo(expectedPointB[0], Assertions.within(1f))
+        Assertions.assertThat(pointBy).isCloseTo(expectedPointB[1], Assertions.within(1f))
     }
 
     @Test
@@ -22,8 +21,8 @@ class TestUtilsTest {
         val pointA = arrayOf(20f, 20f)
         val expectedPointB = arrayOf(20f, 40f)
         val (pointBx, pointBy) = getDestinationPoint(pointA, 20, 90.0)
-        assertThat(pointBx).isCloseTo(expectedPointB[0], within(1f))
-        assertThat(pointBy).isCloseTo(expectedPointB[1], within(1f))
+        Assertions.assertThat(pointBx).isCloseTo(expectedPointB[0], Assertions.within(1f))
+        Assertions.assertThat(pointBy).isCloseTo(expectedPointB[1], Assertions.within(1f))
     }
 
     @Test
@@ -31,8 +30,8 @@ class TestUtilsTest {
         val pointA = arrayOf(20f, 20f)
         val expectedPointB = arrayOf(34.14f, 34.14f)
         val (pointBx, pointBy) = getDestinationPoint(pointA, 20, 45.0)
-        assertThat(pointBx).isCloseTo(expectedPointB[0], within(1f))
-        assertThat(pointBy).isCloseTo(expectedPointB[1], within(1f))
+        Assertions.assertThat(pointBx).isCloseTo(expectedPointB[0], Assertions.within(1f))
+        Assertions.assertThat(pointBy).isCloseTo(expectedPointB[1], Assertions.within(1f))
     }
 
     @Test
@@ -40,7 +39,7 @@ class TestUtilsTest {
         val pointA = arrayOf(20f, 20f)
         val expectedPointB = arrayOf(5.86f, 5.86f)
         val (pointBx, pointBy) = getDestinationPoint(pointA, 20, 225.0)
-        assertThat(pointBx).isCloseTo(expectedPointB[0], within(1f))
-        assertThat(pointBy).isCloseTo(expectedPointB[1], within(1f))
+        Assertions.assertThat(pointBx).isCloseTo(expectedPointB[0], Assertions.within(1f))
+        Assertions.assertThat(pointBy).isCloseTo(expectedPointB[1], Assertions.within(1f))
     }
 }
